[quant][graphmode][fx] Add default_reuse_input_qconfig for ops like reshape
Summary:
This PR added default_reuse_input_qconfig and default_reuse_input_observer which is used
when we want to configure an operator to reuse the observer from input, if the input
Tensor for the operator is not observed, we will not observe the output of this operator either,
if the input Tensor is observed, we will observe the output of the operator with the same observer.

e.g.

```
x1 = x0.reshape()
```
if reshape is configured with default_reuse_input_qconfig:
1. if x0 is observed with observer_0, we'll observe x1 with the same observer instance
2. if x0 is not observed, we won't observe x1 either

Test Plan:
```
python test/test_quantization.py TestQuantizeFx.test_reuse_input_qconfig
```
Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 54df42165eb32b9da3b619fb23098343d13471d2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69249